backend/requirements.txt
Django==3.1.4
psycopg2>=2.7,<3.0
django-cors-headers==3.6.0
djangorestframework==3.12.2
Pillow==8.0.1